Work around cuda parameter space limit
Specifically, with previous commit we observe:

gpu_constructor.cpp(102): Error: Formal parameter space overflowed (5928 bytes required, max 4096 bytes allowed) in function _ZN3gpu26instantiateDeviceChemistryEP10GasMixture14ChemistryInputPv

which is happening b/c we increased gpudata::MAXREACTIONS, making
struct ChemistryInput too large for cuda to pass.  In general, need to
refactor how sizing and passing are handled in the gpu path to avoid
this kind of thing, but for now, we make the ChemistryInput struct
more efficient by using short int for the stoichimetric coefficients
(rather than double).

+++ src/dataStructures.hpp
@@ -632,8 +632,14 @@ struct ChemistryInput {
   int numReactions;
   double reactionEnergies[gpudata::MAXREACTIONS];
   bool detailedBalance[gpudata::MAXREACTIONS];
-  double reactantStoich[gpudata::MAXSPECIES * gpudata::MAXREACTIONS];
-  double productStoich[gpudata::MAXSPECIES * gpudata::MAXREACTIONS];
+  // NOTE(trevilo): reactantStoich and productStoich used to be
+  // double.  Changed to short int to conserve space to satisfy cuda
+  // parameter space limit, which was breached after increasing
+  // MAXREACTIONS.  Causes no issue now, since our stoichiometric
+  // coefficients are alwas small integers, but if ever need more
+  // generality for some reason, then it will have to be implemented.
+  short int reactantStoich[gpudata::MAXSPECIES * gpudata::MAXREACTIONS];
+  short int productStoich[gpudata::MAXSPECIES * gpudata::MAXREACTIONS];
   ReactionModel reactionModels[gpudata::MAXREACTIONS];
   double equilibriumConstantParams[gpudata::MAXCHEMPARAMS * gpudata::MAXREACTIONS];
   ReactionInput reactionInputs[gpudata::MAXREACTIONS];
